@ Build: migrate from v140/Win8.1SDK to v145/Win10SDK (VS2026)
- Update PlatformToolset v140->v145, Windows SDK 8.1->10.0.26100.0
- Disable ConformanceMode, add ForceInclude for string/cmath/stdexcept
- Disable WholeProgramOptimization to avoid LNK1257 LTCG failure
- Add NoLifeNx/nlnx/nx.cpp to project for missing nl::nx::* globals
- Fix narrowing conversions (C2398) in Point.h, Range.h, DrawArgument.h
- Fix missing includes: <string> in Job.h/Equip.h, <stdexcept> in EnumMap.h, <algorithm> in Color.cpp
- Fix narrowing in Combat.cpp, MapMobs.cpp, FootholdTree.cpp
@

+++ Gameplay/MapleMap/MapMobs.cpp
@@ -131,7 +131,7 @@ namespace ms
 		Rectangle<int16_t> player_rect = {
 			horizontal.smaller(),
 			horizontal.greater(),
-			vertical.smaller() - 50,
+			static_cast<int16_t>(vertical.smaller() - 50),
 			vertical.greater()
 		};
